+++ tools/ldb_tool.cc
@@ -30,6 +30,10 @@ public:
         " : Values are input/output as hex\n");
     ret.append("  --" + LDBCommand::ARG_HEX +
         " : Both keys and values are input/output as hex\n");
+    ret.append(
+        "  --" + LDBCommand::ARG_CF_NAME +
+        " : name of the column family to operate on. default: default column "
+        "family\n");
     ret.append("\n");
 
     ret.append("The following optional parameters control the database "
@@ -77,15 +81,16 @@ public:
     fprintf(stderr, "%s\n", ret.c_str());
   }
 
-  static void RunCommand(int argc, char** argv, Options options,
-                         const LDBOptions& ldb_options) {
+  static void RunCommand(
+      int argc, char** argv, Options options, const LDBOptions& ldb_options,
+      const std::vector<ColumnFamilyDescriptor>* column_families) {
     if (argc <= 2) {
       PrintHelp(argv[0]);
       exit(1);
     }
 
-    LDBCommand* cmdObj = LDBCommand::InitFromCmdLineArgs(argc, argv, options,
-                                                         ldb_options);
+    LDBCommand* cmdObj = LDBCommand::InitFromCmdLineArgs(
+        argc, argv, options, ldb_options, column_families);
     if (cmdObj == nullptr) {
       fprintf(stderr, "Unknown command\n");
       PrintHelp(argv[0]);
@@ -106,10 +111,11 @@ public:
 
 };
 
-
 void LDBTool::Run(int argc, char** argv, Options options,
-                  const LDBOptions& ldb_options) {
-  LDBCommandRunner::RunCommand(argc, argv, options, ldb_options);
+                  const LDBOptions& ldb_options,
+                  const std::vector<ColumnFamilyDescriptor>* column_families) {
+  LDBCommandRunner::RunCommand(argc, argv, options, ldb_options,
+                               column_families);
 }
 } // namespace rocksdb